Modify Repositories

+++ src/main/java/srl/narrel/demo/repositories/AppartamentoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import srl.narrel.demo.models.AppartamentoModel;
 
-public interface AppartamentoRepository extends JpaRepository<AppartamentoModel, Long> {
+public interface AppartamentoRepository extends JpaRepository<AppartamentoModel, Integer> {
  
 }
